Revert "Pull up LayerLwIP::ScheduleLambda to SystemLayer (#11187)"
This reverts commit 16cd474e9068e3184865603e8812849b1cc0644c.
